+++ tests/test_mx_adapter.py
@@ -957,6 +957,7 @@ class TestMxAdapter:
                 loras=[lora],
                 lora_strengths=[0.5],
                 lora_mode="merge",
+                stochastic_rounding=False,
                 stream_config=StreamConfig(num_resident_blocks=1, num_prefetch_blocks=0),
             ) as active:
                 merged = active.blocks[0].weight.data